requirements.txt
numpy==2.3.4
Pillow==12.0.0
torch==2.5.1
transformers==4.52.4
